+++ packages/analytics-js-common/src/utilities/json.ts
@@ -1,12 +1,17 @@
 import type { ILogger } from '../types/Logger';
 import type { Nullable } from '../types/Nullable';
-import { isNull, isNullOrUndefined } from './checks';
+import { isBigInt, isNull, isNullOrUndefined } from './checks';
 import {
+  BAD_DATA_WARNING,
   CIRCULAR_REFERENCE_WARNING,
   JSON_STRINGIFY_WARNING,
 } from '../constants/logMessages';
+import { isObjectLiteralAndNotNull } from './object';
 
 const JSON_STRINGIFY = 'JSONStringify';
+const JSON_UTIL = 'JSON';
+const BIG_INT_PLACEHOLDER = '[BigInt]';
+const CIRCULAR_REFERENCE_PLACEHOLDER = '[Circular Reference]';
 
 const getCircularReplacer = (
   excludeNull?: boolean,
@@ -39,7 +44,7 @@ const getCircularReplacer = (
 
     if (ancestors.includes(value)) {
       logger?.warn(CIRCULAR_REFERENCE_WARNING(JSON_STRINGIFY, key));
-      return '[Circular Reference]';
+      return CIRCULAR_REFERENCE_PLACEHOLDER;
     }
 
     ancestors.push(value);
@@ -89,4 +94,78 @@ const stringifyData = <T = Record<string, any> | any[] | number | string>(
     return value;
   });
 
-export { stringifyWithoutCircular, stringifyData };
+const getReplacer = (logger?: ILogger): ((key: string, value: any) => any) => {
+  const ancestors: any[] = []; // Array to track ancestor objects
+
+  // Using a regular function to use `this` for the parent context
+  return function replacer(key, value): any {
+    if (isBigInt(value)) {
+      logger?.warn(BAD_DATA_WARNING(JSON_UTIL, key));
+      return BIG_INT_PLACEHOLDER; // Replace BigInt values
+    }
+
+    // `this` is the object that value is contained in, i.e., its direct parent.
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore-next-line
+    while (ancestors.length > 0 && ancestors[ancestors.length - 1] !== this) {
+      ancestors.pop(); // Remove ancestors that are no longer part of the chain
+    }
+
+    // Check for circular references (if the value is already in the ancestors)
+    if (ancestors.includes(value)) {
+      logger?.warn(BAD_DATA_WARNING(JSON_UTIL, key));
+      return CIRCULAR_REFERENCE_PLACEHOLDER;
+    }
+
+    // Add current value to ancestors
+    ancestors.push(value);
+
+    return value;
+  };
+};
+
+const traverseWithThis = (obj: any, replacer: (key: string, value: any) => any): any => {
+  // Create a new result object or array
+  const result = Array.isArray(obj) ? [] : {};
+
+  // Traverse object properties or array elements
+  // eslint-disable-next-line no-restricted-syntax
+  for (const key in obj) {
+    if (Object.hasOwnProperty.call(obj, key)) {
+      const value = obj[key];
+
+      // Recursively apply the replacer and traversal
+      const sanitizedValue = replacer.call(obj, key, value);
+
+      // If the value is an object or array, continue traversal
+      if (isObjectLiteralAndNotNull(sanitizedValue) || Array.isArray(sanitizedValue)) {
+        (result as any)[key] = traverseWithThis(sanitizedValue, replacer);
+      } else {
+        (result as any)[key] = sanitizedValue;
+      }
+    }
+  }
+
+  return result;
+};
+
+/**
+ * Recursively traverses an object similar to JSON.stringify,
+ * sanitizing BigInts and circular references
+ * @param value Input object
+ * @param logger Logger instance
+ * @returns Sanitized value
+ */
+const getSanitizedValue = <T = any>(value: T, logger?: ILogger): T => {
+  const replacer = getReplacer(logger);
+
+  // This is needed for registering the first ancestor
+  const newValue = replacer.call(value, '', value);
+
+  if (isObjectLiteralAndNotNull(value) || Array.isArray(value)) {
+    return traverseWithThis(value, replacer);
+  }
+  return newValue;
+};
+
+export { stringifyWithoutCircular, stringifyData, getSanitizedValue };