+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.6.14](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.6.13...@rudderstack/analytics-js-plugins@3.6.14) (2024-11-22)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.10`
+* `@rudderstack/analytics-js` updated to version `3.11.10`
+* `@rudderstack/analytics-js-cookies` updated to version `0.4.12`
+
+### Bug Fixes
+
+* restore data sanitization changes but avoid using api overloads ([d0913ae](https://github.com/rudderlabs/rudder-sdk-js/commit/d0913ae32a8c63def26c081c7570a9960dcd1ebf))
+
 ## [3.6.13](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.6.12...@rudderstack/analytics-js-plugins@3.6.13) (2024-11-22)
 
 ### Dependency Updates