+++ packages/analytics-v1.1/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.48.38](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.48.37...rudder-sdk-js@2.48.38) (2024-11-22)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.14.10`
 ## [2.48.37](https://github.com/rudderlabs/rudder-sdk-js/compare/rudder-sdk-js@2.48.36...rudder-sdk-js@2.48.37) (2024-11-22)
 
 ### Dependency Updates