+++ packages/loading-scripts/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [3.0.54](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.53...@rudderstack/analytics-js-loading-scripts@3.0.54) (2024-11-22)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js` updated to version `3.11.11`
 ## [3.0.53](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-loading-scripts@3.0.52...@rudderstack/analytics-js-loading-scripts@3.0.53) (2024-11-22)
 
 ### Dependency Updates